Fix fit-service

+++ Team_03/02_Source/01_Source_Code/BE/services/fit-service/src/main/java/com/hcmus/fitservice/dto/MealEntryRequestDto.java
@@ -14,5 +14,5 @@ public class MealEntryRequestDto {
 
     private String servingUnit;
 
-    private Integer numberOfServings;
+    private Double numberOfServings;
 }
